requirements.txt
slack-bolt
snowflake-snowpark-python
slack_sdk
pandas
snowflake-connector-python[pandas]